Refactor IvleLoginManager::getUserWithAuthToken()
The logic in getUserWithAuthToken() is too similar to
SmartphoneCardLoginManager::loginToSmartphoneCardSystem().

As a step towards extracting the common logic of both methods, refactor
IvleLoginManager::getUserWithAuthToken(), by extracting the
construction of request and waiting for timeout logic into HttpManager
instead.

That way, getUserWithAuthToken() will only deal with handling the
response given by the IVLE server, instead of having to deal with HTTP
on its own.